requirements.txt
# Core dependencies
pymilvus==2.4.8
neo4j
python-dotenv==1.0.0
python-docx
nltk

# FastAPI and server dependencies
fastapi==0.100.0
uvicorn==0.22.0
pydantic>=2.3.0

# LangChain and related
langchain
langchain-community

# Machine Learning and NLP
numpy

# Utilities
tqdm
setuptools==70.0.0
requests

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Hugging Face
huggingface-hub>=0.19.0
transformers>=4.36.0
accelerate>=0.25.0
torch>=2.0.0
bitsandbytes>=0.41.3

paramiko>=3.4.0  # For SSH communication
asyncssh>=2.13.2  # For async SSH operations
python-multipart  # For file uploads in FastAPI
aiofiles>=23.2.1  # For async file operations
pylt
py-localtunnel